+++ run_tests.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
-# Define the PHPUnit command as a variable
-phpunit_cmd="../../vendor/bin/phpunit --colors --testsuite local_deepler_testsuite"
-# Define the initialisation script
+
+# Check if an argument is provided
+if [ -z "$1" ]; then
+  # Define the PHPUnit command as a variable
+  phpunit_cmd="../../vendor/bin/phpunit --colors --testsuite local_deepler_testsuite"
+else
+  # Define the PHPUnit command with the filter argument
+  phpunit_cmd="../../vendor/bin/phpunit --colors --testsuite local_deepler_testsuite --filter $1"
+fi
+
+# Define the initialization script
 init_phpunit="php ../../admin/tool/phpunit/cli/init.php"
 
 # Run the PHPUnit command and capture the output