+++ tests/local/data/langhelper_test.php
@@ -61,7 +61,7 @@ final class langhelper_test extends advanced_testcase {
         parent::setUp();
         $this->resetAfterTest(true);
         $this->langhelper = new lang_helper();
-        $this->langhelper->init('abcd');
+        $this->langhelper->init(getenv('API_TOKEN'));
     }
 
     /**
@@ -96,4 +96,53 @@ final class langhelper_test extends advanced_testcase {
         $this->assertStringContainsString('<option', $htmloptions);
     }
 
+    /**
+     * @return void
+     * @throws \DeepL\DeepLException
+     * @throws \dml_exception
+     */
+    public function test_settings(): void {
+
+        $key = '';
+        if($this->langhelper->isapikeynoset()){
+            $this->makeenv();
+            $key = getenv('API_TOKEN');
+        }
+        $this->assertIsBool($this->langhelper->init($key));
+    }
+
+    /**
+     * Helper for running test without.
+     *
+     * @return void
+     */
+    private function makeenv() {
+        global $CFG;
+        // Define the path to the .env file
+        $envFilePath = $CFG->dirroot . '/local/deepler/.env';
+
+        // Check if the .env file exists
+        if (file_exists($envFilePath)) {
+            // Read the .env file line by line
+            $lines = file($envFilePath, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+            foreach ($lines as $line) {
+                // Skip comments
+                if (strpos(trim($line), '#') === 0) {
+                    continue;
+                }
+
+                // Parse the environment variable
+                list($name, $value) = explode('=', $line, 2);
+                $name = trim($name);
+                $value = trim($value);
+
+                // Set the environment variable
+                putenv(sprintf('%s=%s', $name, $value));
+                $_ENV[$name] = $value;
+                $_SERVER[$name] = $value;
+            }
+        } else {
+            $this->assertEquals('DEFAULT', getenv('API_TOKEN'));
+        }
+    }
 }